+++ CHANGELOG.md
@@ -2,7 +2,25 @@
 
 Please mark all change in change log and use the issue from GitHub
 
-# Milvus 0.7.0 (TBD)
+# Milvus 0.7.1 (TBD)
+
+## Bug
+-   \#1301 Data in WAL may be accidentally inserted into a new table with the same name.
+-   \#1634 Fix search demo bug in HTTP doc
+-   \#1635 Vectors can be returned by searching after vectors deleted if `cache_insert_data` set true
+
+## Feature
+
+## Improvement
+-   \#1537 Optimize raw vector and uids read/write
+-   \#1546 Move Config.cpp to config directory
+-   \#1547 Rename storage/file to storage/disk and rename classes
+-   \#1548 Move store/Directory to storage/Operation and add FSHandler
+
+## Task
+
+
+# Milvus 0.7.0 (2020-03-11)
 
 ## Bug
 -   \#715 Milvus crash when searching and building index simultaneously using SQ8H
@@ -108,9 +126,7 @@ Please mark all change in change log and use the issue from GitHub
 -   \#1448 General proto api for NNS libraries 
 -   \#1480 Add return code for AVX512 selection
 -   \#1524 Update config "preload_table" description
--   \#1537 Optimize raw vector and uids read/write
 -   \#1544 Update resources name in HTTP module
--   \#1546 Move Config.cpp to config directory
 -   \#1567 Update yaml config description
 
 ## Task